+++ lib/src/levels/_levels.dart
@@ -1,3 +1,4 @@
+export 'controls_menu.dart';
 export 'conversation_level.dart';
 export 'credits_menu.dart';
 export 'custom_menu_level.dart';